Clean up and work on MALA
