--- pism_emulator/models/__init__.py
@@ -1,3 +1,0 @@
-name = "models"
-
-__all__ = ["pdd", "debm"]
