+++ pism_emulator/stats.py
@@ -22,7 +22,7 @@ Stats.
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Iterable, List, Tuple, Sequence, Dict, Any, Optional
+from typing import Any, Dict, Iterable, List, Optional, Sequence, Tuple
 
 import numpy as np
 import pandas as pd
